Address review: privatize force_positive, OceanCorrectorConfig docstring, move utils tests
- Rename force_positive -> _force_positive (internal-only helper)
- Give OceanCorrectorConfig a Parameters docstring covering each field
  (replaces the attribute-docstring on keep_gradient_through_clamps)
- Move force_positive / replace_value_keep_gradient tests from
  test_atmosphere.py into a dedicated test_utils.py
